% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.tpp.stlpp.R
\name{as.tpp.stlpp}
\alias{as.tpp.stlpp}
\title{Convert data to a one-dimensional point pattern}
\usage{
as.tpp.stlpp(X)
}
\arguments{
\item{X}{an object of class \code{\link{stlpp}}}
}
\value{
An object of class tpp.
}
\description{
This function converts an object of class \code{\link{stlpp}} to  class \code{\link{tpp}}.
}
\details{
This function projects the spatio-temporal point pattern X into its corresponding time domain T.
}
\examples{
 
X <- rpoistlpp(10,1,2,easynet)
as.tpp.stlpp(X)

}
\seealso{
\code{\link{as.stlpp}}, \code{\link{lpp}}, \code{\link{as.lpp}}
}
\author{
Mehdi Moradi <m2.moradi@yahoo.com>
}
