\name{print.stlpp}
\alias{print.stlpp}
\title{Methods for space-time point patterns on a linear network.}
\usage{
print.stlpp(x)
}
\description{
Methods to print an object of class \code{\link{stlpp}}.
}
\author{
Mehdi Moradi <m2.moradi@yahoo.com>
}

\seealso{
\code{\link{as.stlpp}}
}
\examples{
X = rpoistlpp(.2,a=0,b=5,L=easynet)
X
}
