% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp.R
\name{tpp}
\alias{tpp}
\title{Create a temporal point pattern}
\usage{
tpp(X,a,b)
}
\arguments{
\item{X}{an object of class \code{\link{numeric}}, \code{\link{integer}} or \code{\link{vector}}}

\item{a}{lower band of the time domain. if not given by the user, it will be the minimum of X}

\item{b}{upper bound of the time domain. if not given by the user, it will be the maximum of X}
}
\value{
An object of class tpp.
}
\description{
Create an object of class \code{\link{tpp}} representing a one-dimensional point pattern.
}
\details{
Create a one-dimensional point pattern.
}
\examples{
 
tpp(runif(10))

}
\seealso{
\code{\link{stlpp}}
}
\author{
Mehdi Moradi <m2.moradi@yahoo.com>
}
